requirements.txt
# Core ML libraries
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Dataset handling
fiftyone>=0.21.0
pandas>=1.3.0

# Image processing and visualization
opencv-python>=4.6.0
pillow>=9.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
tqdm>=4.64.0
numpy>=1.21.0
pyyaml>=6.0
requests>=2.28.0

# Development tools
jupyter>=1.0.0
ipykernel>=6.15.0